+++ R/disastr.api.internal.R
@@ -2,8 +2,9 @@
 .onAttach <-
   function(libname, pkgname) {
     packageStartupMessage("\nThis package may be cited as:")
-    packageStartupMessage('Dworschak, Christoph. 2020. "Disastr.api: Automated Retrieval of Disaster Event Data. R package.')
-    packageStartupMessage('Development version 1.0.0. \n')
+    packageStartupMessage('Dworschak, Christoph. 2021. "Disastr.api: Automated Retrieval ')
+    packageStartupMessage('of Disaster Event Data." R package. CRAN version 1.0.1. \n')
+    packageStartupMessage("For the development version of this package, visit <https://github.com/chris-dworschak/disastr.api/> \n")
   }