+++ NEWS.md
@@ -1,6 +1,23 @@
 
 # Release updates
 
+<!-- Changes for future versions (long term):
+* Built-in functionality to retrieve all available events (argument all.events) through a for loop
+ -->
+
+ 
+## disastr.api version 1.0.1
+
+* Substantial revisions of package documentation and functionality
+* New arguments included in this package version: 
+  + date.start
+  + date.end
+  + disaster
+  + country
+  + status
+  + exact.match
+
+  
 
 ## disastr.api version 1.0.0